feat: porting over lping prefix commands to slash commands
