+++ slashCommands/pings/lpingping.js
@@ -2,6 +2,7 @@ const { SlashCommand } = require("discord-akairo");
 const { Constants, MessageEmbed, MessageActionRow, MessageButton } = require("discord.js");
 const { db } = require("../../models/db");
 const { isTrusted } = require("../../utilities/permissions");
+const { autocomplete } = require("../../slashCommandUtilities/lpingutilities");
 
 class LPingPingCommand extends SlashCommand {
     constructor() {
@@ -107,22 +108,7 @@ class LPingPingCommand extends SlashCommand {
     }
 
     async autocomplete(interaction) {
-        if (!interaction.guild) return interaction.respond([]);
-        const input = interaction.options.getString('pinglist', true).toLowerCase();
-
-        const sql = input.length > 0
-            ? 'SELECT `name`, `guildID` FROM `Ping` WHERE `guildID` = ? AND `name` LIKE ? LIMIT 10'
-            : 'SELECT `name`, `guildID` FROM `Ping` WHERE `guildID` = ? ORDER BY `name` LIMIT 10';
-
-        const values = input.length > 0 ? [interaction.guild.id, `${input}%`] : [interaction.guild.id];
-        const [row] = await db.promise().query(sql, values);
-
-        const result = row.map((x) => ({
-            name: x.name,
-            value: x.name,
-        }));
-
-        await interaction.respond(result);
+        await autocomplete(interaction);
     }
 }
